01/11/2024

+++ 233385_C1_ProyectoWeb/src/app/Login/login/login.component.ts
@@ -10,16 +10,33 @@ import Swal from 'sweetalert2';
 export class LoginComponent {
   username: string = '';
   password: string = '';
+  private adminUsers: { username: string; password: string }[] = [
+    { username: 'admin1', password: 'admin123' },
+    { username: 'admin2', password: 'admin456' }
+  ];
 
   constructor(private router: Router) {}
 
   onSubmit() {
     const users = JSON.parse(localStorage.getItem('users') || '[]');
-    const user = users.find((u: any) => u.username === this.username && u.password === this.password);
-    
+    let user = users.find((u: any) => u.username === this.username && u.password === this.password);
+    if (!user) {
+      user = this.adminUsers.find(admin => admin.username === this.username && admin.password === this.password);
+      if (user) {
+        Swal.fire({
+          icon: 'info',
+          title: '¡Atención!',
+          text: 'Eres un usuario administrador.',
+          confirmButtonText: 'Aceptar'
+        });
+      }
+    }
     if (user) {
       user.iniciosesion = true;
-      localStorage.setItem('users', JSON.stringify(users));
+      if (users.some((u: any) => u.username === user.username)) {
+        localStorage.setItem('users', JSON.stringify(users));
+      }
+      localStorage.setItem('loggedInUser', JSON.stringify(user)); 
       this.router.navigate(['/home']);
     } else {
       Swal.fire({
@@ -29,5 +46,5 @@ export class LoginComponent {
         confirmButtonText: 'Aceptar'
       });
     }
-  }
+  }  
 }
